+++ src/encoding.js
@@ -15,8 +15,8 @@ export function bitWidth(value) {
  *
  * If length is zero, then read int32 length at the start.
  *
- * @typedef {import("./types.d.ts").DataReader} DataReader
- * @typedef {import("./types.d.ts").DecodedArray} DecodedArray
+ * @typedef {import("../src/types.d.ts").DataReader} DataReader
+ * @typedef {import("../src/types.d.ts").DecodedArray} DecodedArray
  * @param {DataReader} reader
  * @param {number} width - width of each bit-packed group
  * @param {number} length - length of the encoded data
@@ -117,7 +117,7 @@ function readBitPacked(reader, header, bitWidth, output, seen) {
 }
 
 /**
- * @typedef {import("./types.d.ts").ParquetType} ParquetType
+ * @typedef {import("../src/types.d.ts").ParquetType} ParquetType
  * @param {DataReader} reader
  * @param {number} count
  * @param {ParquetType} type
