+++ src/header.js
@@ -4,8 +4,8 @@ import { deserializeTCompactProtocol } from './thrift.js'
 /**
  * Read parquet header from a buffer.
  *
- * @typedef {import("./types.d.ts").DataReader} DataReader
- * @typedef {import("./types.d.ts").PageHeader} PageHeader
+ * @typedef {import("../src/types.d.ts").DataReader} DataReader
+ * @typedef {import("../src/types.d.ts").PageHeader} PageHeader
  * @param {DataReader} reader - parquet file reader
  * @returns {PageHeader} metadata object and bytes read
  */
