+++ src/metadata.js
@@ -23,9 +23,9 @@ import { deserializeTCompactProtocol } from './thrift.js'
  * This ensures that we either make one 512kb initial request for the metadata,
  * or a second request for up to the metadata size.
  *
- * @typedef {import("./types.d.ts").AsyncBuffer} AsyncBuffer
- * @typedef {import("./types.d.ts").FileMetaData} FileMetaData
- * @typedef {import("./types.d.ts").SchemaElement} SchemaElement
+ * @typedef {import("../src/types.d.ts").AsyncBuffer} AsyncBuffer
+ * @typedef {import("../src/types.d.ts").FileMetaData} FileMetaData
+ * @typedef {import("../src/types.d.ts").SchemaElement} SchemaElement
  * @param {AsyncBuffer} asyncBuffer parquet file contents
  * @param {number} initialFetchSize initial fetch size in bytes
  * @returns {Promise<FileMetaData>} parquet metadata object
@@ -190,7 +190,7 @@ export function parquetMetadata(arrayBuffer) {
  * Return a tree of schema elements from parquet metadata.
  *
  * @param {FileMetaData} metadata parquet metadata object
- * @returns {import("./types.d.ts").SchemaTree} tree of schema elements
+ * @returns {import("../src/types.d.ts").SchemaTree} tree of schema elements
  */
 export function parquetSchema(metadata) {
   return getSchemaPath(metadata.schema, [])[0]
@@ -198,7 +198,7 @@ export function parquetSchema(metadata) {
 
 /**
  * @param {any} logicalType
- * @returns {import("./types.d.ts").LogicalType | undefined}
+ * @returns {import("../src/types.d.ts").LogicalType | undefined}
  */
 function logicalType(logicalType) {
   if (logicalType?.field_1) return { type: 'STRING' }
@@ -236,7 +236,7 @@ function logicalType(logicalType) {
 
 /**
  * @param {any} unit
- * @returns {import("./types.d.ts").TimeUnit}
+ * @returns {import("../src/types.d.ts").TimeUnit}
  */
 function timeUnit(unit) {
   if (unit.field_1) return 'MILLIS'
@@ -250,7 +250,7 @@ function timeUnit(unit) {
  *
  * @param {any} stats
  * @param {SchemaElement} schema
- * @returns {import("./types.d.ts").Statistics}
+ * @returns {import("../src/types.d.ts").Statistics}
  */
 function convertStats(stats, schema) {
   return stats && {
@@ -268,7 +268,7 @@ function convertStats(stats, schema) {
 /**
  * @param {Uint8Array | undefined} value
  * @param {SchemaElement} schema
- * @returns {import('./types.d.ts').MinMaxType | undefined}
+ * @returns {import('../src/types.d.ts').MinMaxType | undefined}
  */
 export function convertMetadata(value, schema) {
   const { type, converted_type, logical_type } = schema
