+++ src/plain.js
@@ -1,9 +1,9 @@
 /**
  * Read `count` values of the given type from the reader.view.
  *
- * @typedef {import("./types.d.ts").DataReader} DataReader
- * @typedef {import("./types.d.ts").DecodedArray} DecodedArray
- * @typedef {import("./types.d.ts").ParquetType} ParquetType
+ * @typedef {import("../src/types.d.ts").DataReader} DataReader
+ * @typedef {import("../src/types.d.ts").DecodedArray} DecodedArray
+ * @typedef {import("../src/types.d.ts").ParquetType} ParquetType
  * @param {DataReader} reader - buffer to read data from
  * @param {ParquetType} type - parquet type of the data
  * @param {number} count - number of values to read
